server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.9.2
requests==2.32.3
supabase==2.5.0
python-dotenv==1.0.1